+++ blueprint/iam-better-auth/mikro-orm.config.ts
@@ -1,16 +1,17 @@
 import { number, schemaValidator, string } from '@forklaunch/blueprint-core';
 import {
+  ComplianceEventSubscriber,
+  FieldEncryptor
+} from '@forklaunch/core/persistence';
+import {
   createConfigInjector,
   getEnvVar,
   Lifetime
 } from '@forklaunch/core/services';
 import { Platform, TextType, Type } from '@mikro-orm/core';
 import { Migrator } from '@mikro-orm/migrations';
-import { ComplianceEventSubscriber } from '@forklaunch/core/persistence';
-import { FieldEncryptor } from '@forklaunch/core/encryption';
 import { defineConfig } from '@mikro-orm/postgresql';
 import dotenv from 'dotenv';
-
 import * as entities from './persistence/entities';
 
 dotenv.config({ path: getEnvVar('DOTENV_FILE_PATH') });
@@ -45,19 +46,25 @@ const configInjector = createConfigInjector(schemaValidator, {
     lifetime: Lifetime.Singleton,
     type: string,
     value: getEnvVar('NODE_ENV')
+  },
+  ENCRYPTION_KEY: {
+    lifetime: Lifetime.Singleton,
+    type: string,
+    value: getEnvVar('ENCRYPTION_KEY')
   }
 });
 
 export const validConfigInjector = configInjector.validateConfigSingletons(
   getEnvVar('DOTENV_FILE_PATH') ?? '.env'
 );
+const tokens = validConfigInjector.tokens();
 
 const mikroOrmOptionsConfig = defineConfig({
-  dbName: validConfigInjector.resolve('DB_NAME'),
-  host: validConfigInjector.resolve('DB_HOST'),
-  user: validConfigInjector.resolve('DB_USER'),
-  password: validConfigInjector.resolve('DB_PASSWORD'),
-  port: validConfigInjector.resolve('DB_PORT'),
+  dbName: validConfigInjector.resolve(tokens.DB_NAME),
+  host: validConfigInjector.resolve(tokens.DB_HOST),
+  user: validConfigInjector.resolve(tokens.DB_USER),
+  password: validConfigInjector.resolve(tokens.DB_PASSWORD),
+  port: validConfigInjector.resolve(tokens.DB_PORT),
   entities: Object.values(entities),
   subscribers: [
     new ComplianceEventSubscriber(